feat: Add react-simple-wysiwyg package
Add react-simple-wysiwyg package to the project dependencies in package.json and package-lock.json. This package will be used for implementing a rich text editor in the application.

Also, create a new component RichTextEditor.jsx that uses the react-simple-wysiwyg package to render a toolbar and editor for formatting text.

Lastly, add a new form component Experience.jsx that includes the RichTextEditor component for capturing work summary in the professional experience section.

Closes #123